Add file logging support.
Configure the logging so that it logs to disk as well as logging to
console. Additionally, this change removes the dependency on
coloredlogs, though that may be reintroduced later for the console
output. There may be some errors when multiple processes try to log to
the same file, but it seems to work well enough for now.

Introduces dependency on pyYAML.